refactor(web): promote callServerFn and extract shared UI primitives
- callServerFn (optional cause mapper) replaces ~12 inline Effect error-folding copies; changeRole's re-read no longer escapes as unhandled rejection
- ConfirmButton primitive replaces three hand-rolled arm/confirm/cancel flows
- auth-audit split into shared/lifecycle/admin modules with direct imports
- Single EMAIL_PATTERN, badge status mappers, and URL-based TOTP secret parsing
- Two-factor panel handlers collapse into one runAction helper
